requirements.txt
torch>=2.2.0
torchvision>=0.17.0
tensordict>=0.6.1
torchrl>=0.8.0
vmas@git+https://github.com/facebookresearch/vmas.git
matplotlib>=3.7.0
tqdm>=4.66.0
pyvirtualdisplay>=3.0
Pillow>=10.0.0
ipython>=8.0.0
